Move user interactions manager, window instance holder and windows helper into services/windows
