+++ kiosk/src/js/application_updater.js
@@ -4,7 +4,7 @@ const CurrentScreenSettingsManager = require('./current_screen_settings_manager'
 const cron = require('node-cron');
 const NotificationListener = require('./services/network/notification_listener');
 const PropertiesLoader = require('./helpers/properties_load_helper');
-const UserInteractionsManager = require('./user_interactions_manager');
+const UserInteractionsManager = require('./services/windows/user_interactions_manager');
 const NetworkErrorsHandlingService = require('./services/error-handlers/network_errors_handling_service');
 const AutoupdateScheduleWatcher = require('./services/data/autoupdate_schedule_watcher');
 const ConnectionStatusService = require('./services/network/connection_status_service');
