+++ kiosk/src/js/services/windows/user_interactions_manager.js
