+++ kiosk/src/js/services/windows/window_instance_holder.js
