+++ kiosk/src/main.js
@@ -4,7 +4,7 @@ const ApplicationUpdater = require('./js/application_updater');
 const app = electron.app;
 
 const CurrentScreenSettingsManager = require('./js/current_screen_settings_manager');
-const WindowsHelper = require('./js/helpers/windows_helper');
+const WindowsHelper = require('./js/services/windows/windows_helper');
 const SettingsLoadJobManager = require('./js/helpers/settings_load_job_manager');
 const Logger = require('./js/services/logger/logger');
 const StorageManager = require('./js/helpers/storage_manager');
